Implement outline for computing distances
All set for implementing the 3D distance computation
function. Currently only CSV-BULDINGS-MULTIPLE outputs
the RMSE value, which is simple the the nr. LAS points
in the footprint for now.

+ Modify Map3d::add_las_file() to include a switch
for 'elevation' or 'distance' computation
+ Implement Building::add_point_distance() with a
placeholder for the real distance computing function
+ Add placeholders for all TopoFeature classes for
add_point_distance()
+ Modify Map3d::get_csv_buildings_multiple_heights()
so that it writes the RMSE values