Nasty RMSE bug fixed thanks to @kenohori
The problem was that the cgal_tris was a
local variable in TopoFeature::create_triangle_trees() and the AABB
only keeps a pointer to the triangles and stores only the bounding
boxes. Thus when the create_triangle_trees() function exits, cgal_tris
is destroyed and when I try to compute the distances in a separate
function, the AABB tree might point to non-existing triangles. The
unpredictable behaviour comes from the fact that distances can be
computed as long as memory location of the triangles in cgal_tris does
not change.